% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/some.R
\name{some}
\alias{some}
\title{Calculate the sum of vector elements}
\usage{
some(x)
}
\arguments{
\item{x}{A vector that needs a sum}
}
\value{
The sum of vector elements
}
\description{
Calculate the sum of vector elements
}
\examples{
some(rep(1,3))

}
